requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/damnever/pigar

PyYAML == 5.1.2
autopep8
